@@ -3,16 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "@babel/core": "^7.18.10",
+    "@babel/plugin-syntax-flow": "^7.18.6",
+    "@babel/plugin-transform-react-jsx": "^7.18.10",
+    "nth-check": "^2.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-is": "^18.2.0",
     "react-pdf": "^5.7.2",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "react-scroll": "^1.8.7",
     "styled-components": "^5.3.5",
-    "typescript": "^4.6.4",
-    "web-vitals": "^2.1.4"
+    "terser": "^5.14.2",
+    "typescript": "^4.7.4",
+    "webpack": "^5.74.0"
   },
   "scripts": {
+    "audit": "yarn install --check-files",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
@@ -38,16 +45,13 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
-    "@testing-library/user-event": "^14.2.0",
-    "@types/jest": "^27.5.1",
-    "@types/node": "^17.0.32",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.3",
+    "@types/node": "^18.7.8",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
     "@types/react-pdf": "^5.7.2",
-    "@types/react-scroll": "^1.8.3",
-    "@types/styled-components": "^5.1.25",
-    "gh-pages": "^4.0.0"
+    "@types/react-scroll": "^1.8.4",
+    "@types/styled-components": "^5.1.26",
+    "gh-pages": "^4.0.0",
+    "yarn-upgrade-all": "^0.7.1"
   }
 }
