@@ -1,45 +1,47 @@
 import { ReactNode } from "react";
 import styled from "styled-components";
-import { NAV_BAR_HEIGHT } from "./";
 import ViewHeader, { HEADER_HEIGHT } from "./ViewHeader";
-
-/** Constants */
-
-export const FADE_IN_TIME = 0.45;
 
 /** Types */
 
 type Props = {
+  /** Centers content in the body of the view. Default: true */
+  center?: boolean;
   /** Text displayed on the header of the page. */
   headerText?: string;
   /** Padding applied to the X axis of the page. */
   paddingX?: number;
-  /** Centers content in the body of the view. Default: true */
-  center?: boolean;
+  /** Passes in inline styling to the page component. */
+  style?: React.CSSProperties;
   /** Passes the name of the view as the ID. */
   viewId: string;
   children?: ReactNode;
 };
-type BodyStyleType = {
-  paddingX?: number;
-  center?: boolean;
-  minHeight: number;
-};
+type BodyStyleType = Pick<Props, "center" | "paddingX"> & { minHeight: number };
 
 /** Base view component. */
-const View = ({ viewId, headerText, children, ...passThrough }: Props) => {
+const View = ({
+  viewId,
+  headerText,
+  children,
+  style,
+  ...passThrough
+}: Props) => {
   const windowHeight = window.innerHeight;
   const heightCalc =
     headerText && window.innerWidth > 600
-      ? windowHeight - (NAV_BAR_HEIGHT + HEADER_HEIGHT)
-      : windowHeight - NAV_BAR_HEIGHT;
+      ? windowHeight - HEADER_HEIGHT
+      : windowHeight;
 
   return (
-    <ViewStyle id={viewId}>
-      {headerText && window.innerWidth > 600 && <ViewHeader text={headerText} />}
+    <ViewStyle id={viewId} style={style}>
+      {headerText && window.innerWidth > 600 && (
+        <ViewHeader text={headerText} />
+      )}
       <BodyStyle minHeight={heightCalc} {...passThrough}>
         {children}
       </BodyStyle>
+      <div style={{}}></div>
     </ViewStyle>
   );
 };
@@ -47,40 +49,28 @@ const View = ({ viewId, headerText, children, ...passThrough }: Props) => {
 /** Styles */
 
 const ViewStyle = styled.div`
-  width: 100%;
+  box-sizing: border-box;
   display: flex;
   flex-direction: column;
-  box-sizing: border-box;
+  width: 100%;
   z-index: 2;
+
   /* Invisible, to show the background image. */
   background-color: rgba(245, 245, 245, 0);
 `;
 const BodyStyle = styled.div<BodyStyleType>`
-  min-height: ${({ minHeight }) => `${minHeight}px`};
-  height: auto;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
   align-items: ${({ center }) => {
     if (center === undefined) center = true;
     return center ? "center" : undefined;
   }};
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  height: auto;
+  min-height: ${({ minHeight }) => `${minHeight}px`};
   padding: 24px ${({ paddingX }) => (paddingX ? `${paddingX}px` : "0px")} 0px
     ${({ paddingX }) => (paddingX ? `${paddingX}px` : "0px")};
-
-  /* Disabled for conversion to single page app. */
-  /* opacity: 0;
-  animation: fadeIn ${FADE_IN_TIME}s linear 10ms forwards;
-  transition: linear;
-  @keyframes fadeIn {
-    from {
-      opacity: 0;
-    }
-    to {
-      opacity: 1;
-    }
-  } */
+  width: 100%;
 `;
 
 /** Exports */
