@@ -1,11 +1,8 @@
 export { default as Divider } from "./Divider";
+export { default as FlipCard } from "./FlipCard";
 export { default as Grid } from "./Grid";
-export { default as ViewHeader } from "./ViewHeader";
-export { default as NavBar, NAV_BAR_HEIGHT } from "./Navigation";
+export { default as NewNavV } from "./NewNavigationV";
 export { default as PDF } from "./PDF";
 export { default as Text } from "./Text";
-export { default as View, FADE_IN_TIME } from "./View";
-export { default as FlipCard } from "./FlipCard"
-
-export { default as NewNav } from "./NewNavigation"
-export { default as NewNavV } from "./NewNavigationV"
+export { default as View } from "./View";
+export { default as ViewHeader } from "./ViewHeader";
